django_backend/pyproject.toml
[tool.poetry]
name = "django-backend"
version = "0.1.0"
description = ""
authors = ["masato-210831 <b86500b@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9.13"
django = "3.0.7"
djangorestframework = "3.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
